Add archived prototypes file
This file showcases the many attempts that were made in the mocking stage to create a workable artifact. A "Museum of failure", if you will. Whilst this repository is quite young, the work behind it all has been ongoing for many weeks. MechSync was not built in a day, but many failed implementations were